# api_docs/openapi.yaml
openapi: 3.0.0
info:
  title: LocalGPT Vision API
  version: 1.0.0
  description: API for LocalGPT Vision, a Django-based backend for AI chat and vision applications
servers:
  - url: http://localhost:8000/api
    description: Development server
paths:
...
  /models/files/{file_id}/:
    get:
      summary: Get file details
      description: Get details of a specific uploaded file
      tags:
        - Files
      security:
        - BearerAuth: []
      parameters:
        - name: file_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: File details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadedFile'
    delete:
      summary: Delete a file
      description: Delete a specific uploaded file
      tags:
        - Files
      security:
        - BearerAuth: []
      parameters:
        - name: file_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: File deleted
...
components:
  schemas:
...
    UploadedFile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user:
          type: string
          format: uuid
        file:
          type: string
          description: File URL
        file_type:
          type: string
          enum: [document, image, video, audio]
        upload_date:
          type: string
          format: date-time
        processed:
          type: boolean
        processing_results:
          type: object
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT 